% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{package}
\name{envsocty3LT3-package}
\alias{envsocty3LT3}
\alias{envsocty3LT3-package}
\title{envsocty3LT3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/envsocty3LT3}
  \item Report bugs at \url{https://github.com/paezha/envsocty3LT3/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
