% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_demo_taz}
\alias{hamilton_demo_taz}
\title{Socio-demographic information in the Hamilton CMA.}
\format{
A data frame with 297 rows and 15 variables:
\describe{
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{Population}{Population in traffic analysis zone}
\item{Worked_in_2010_Full_time}{Number of peple in traffic analysis zones that had were employed full time and worked away from home in 2010}
\item{Worked_in_2010_Part_time}{Number of peple in traffic analysis zones that had were employed part time and worked away from home in 2010}
\item{Worked_at_home}{Number of peple in traffic analysis zones that worked from home in 2010}
\item{Pop_Density}{Population density of traffic analysis zone in persons/sq.km}
\item{Median_Age}{Median age of the population in traffic analysis zone}
\item{Family_Size_2}{Number of families of size 2 in the traffic analysis zone}
\item{Family_Size_3}{Number of families of size 3 in the traffic analysis zone}
\item{Family_Size_4}{Number of families of size 4 in the traffic analysis zone}
\item{Family_Size_5_more}{Number of families of size 5 or more in the traffic analysis zone}
\item{Median_income}{Median income of households in traffic analysis zone}
\item{Average_income}{Average income of households in traffic analysis zone}
\item{Employment_rate}{Employment rate in traffic analysis zone}
\item{Unemployment_rate}{Unemployment rate in traffic analysis zone}
\item{Median_commuting_duration}{Median duration in minutes of commute to work trips that originated in traffic analysis zone}
}
}
\source{
Transportation Tomorrow Survey (http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction)
}
\usage{
data(hamilton_demo_taz)
}
\description{
A table with socio-demographicinformation from the Transportation Tomorrow Survey
by traffic analysis zone. These data are for 2016.
}
\examples{
 data(hamilton_demo_taz)
 summary(hamilton_demo_taz)
}
\keyword{datasets}
\keyword{transportation}
