% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_graph}
\alias{hamilton_graph}
\title{Road network in and around the Hamilton CMA.}
\format{
A large tbl_graph object consisting of two tables, one for nodes and one for edges (links)

edges is simple features object with 35,142 rows and 6 variables:
\describe{
\item{from}{Unique identifier of origin node in the network}
\item{to}{Unique identifier of destination node in the network}
\item{highway}{OpenStreetMap classification of the link. For defintions, see: \url{https://wiki.openstreetmap.org/wiki/Key:highway}}
\item{geometry}{Geometry column}
\item{EdgeID}{Unique identifier of the edge}
\item{length}{Length of edge in meters}
}

nodes is simple features object with 25,816 rows and 3 variables. The geometry is :
\describe{
\item{nodeID}{Unique identifier of node}
\item{geometry}{Geometry column}
\item{GTA06}{}
}
}
\source{
OpenStreetMaps (\url{https://www.openstreetmap.org/#map=3/71.34/-96.82})
}
\usage{
data(hamilton_graph)
}
\description{
A tbl_graph object with the road network in and around the Hamilton CMA. The network was
obtained from OpenStreetMap and includes highway links of type: motorway, motorway_link,
primary, primary_link, residential, secondary, secondary_link, tertiary, and tertiary_link.
The tbl_graph object consists of two tables: one table with the edges of the network and
and one table with the nodes of the network. The object is projected as UTM Zone 17N.
}
\examples{
 data(hamilton_graph)
 summary(hamilton_graph)
}
\keyword{datasets}
\keyword{transportation}
