% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_neighborhoods}
\alias{hamilton_neighborhoods}
\title{Hamilton neighborhoods.}
\format{
A simple features object with 137 rows and 4 variables:
\describe{
\item{OBJECTID}{Unique identifier for the neighborhood polygon}
\item{Name}{Name of neighborhood}
\item{Number}{Neighborhood number}
\item{geometry}{Geometry column}
}
}
\source{
OpenStreetMaps (\url{https://www.openstreetmap.org/#map=3/71.34/-96.82})
}
\usage{
data(hamilton_neighborhoods)
}
\description{
A table with traditional neighborhoods in the City of Hamilton. These data were
obtained from Hamilton Open Data. The object is projected as UTM Zone 17N.
}
\examples{
 data(hamilton_neighborhoods)
 summary(hamilton_neighborhoods)
}
\keyword{datasets}
\keyword{land}
\keyword{transportation}
\keyword{uses}
