% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_sustenance}
\alias{hamilton_sustenance}
\title{Establishments in and around the Hamilton CMA that serve food.}
\format{
A simple features object with 3,772 rows and 5 variables:
\describe{
\item{osm_id}{Unique identifier for the point in OpenStreetMap}
\item{name}{Name of establishment (when available)}
\item{amenity}{Amenity category (when available). See: \url{https://wiki.openstreetmap.org/wiki/Key:amenity}}
\item{type}{Type of establishment: Bar, Cafe, Fast Food, Food Court, Pub, Restaurant}
\item{geometry}{Geometry column}
}
}
\source{
OpenStreetMaps (\url{https://www.openstreetmap.org/#map=3/71.34/-96.82})
}
\usage{
data(hamilton_sustenance)
}
\description{
A table with establishments that serve food in and around the Hamilton CMA. These data were
obtained from OpenStreetMap and includes the following types of establishments: Bar, Cafe,
Fast Food, Food Court, Pub, and Restaurant. The object is projected as UTM Zone 17N.
}
\examples{
 data(hamilton_sustenance)
 summary(hamilton_sustenance)
}
\keyword{datasets}
\keyword{land}
\keyword{transportation}
\keyword{uses}
