% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_taz_employment}
\alias{hamilton_taz_employment}
\title{Employment statistics by place of employment for traffic analysis zones in the Hamilton CMA.}
\format{
A simple features object with 297 rows and 5 variables:
\describe{
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{Jobs_Manufacturing_Construction_Trades}{Number of jobs in the traffic analysis zone of type Manufacturing/Construction/Trades}
\item{Jobs_Retail}{Number of jobs in the traffic analysis zone of type Retail, Sales and Service}
\item{Jobs_Office_Clerical}{Number of jobs in the traffic analysis zone of type Manufacturing/Construction/Trades}
\item{Jobs_Professional}{Number of jobs in the traffic analysis zone of type Professional/Management/Technical}
}
}
\source{
Transportation Tomorrow Survey (\url{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction})

Data Guide available (\url{http://dmg.utoronto.ca/pdf/tts/2016/2016TTS_DataGuide.pdf})
}
\usage{
data(hamilton_taz_employment)
}
\description{
A table with statistics about jobs by traffic analysis zones in the Hamilton CMA.
The zoning system is GTA06, and employment statistics are number of jobs in the
traffic analysis zone. This file is a portion of the system used by the Transportation
Tomorrow Survey in the Greater Toronto and Hamilton Area.
}
\examples{
 data(hamilton_taz_employment)
 summary(hamilton_taz_employment)
}
\keyword{datasets}
\keyword{transportation}
