% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_taz_ph}
\alias{hamilton_taz_ph}
\title{Person and households statistics for traffic analysis zones in the Hamilton CMA.}
\format{
A simple features object with 297 rows and 20 variables:
\describe{
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{population}{Number of people in the traffic analysis zone by place of residence}
\item{drivers}{Number of drivers (people who have a driver license) in the traffic analysis zone by place of residence}
\item{vehicles}{Number of privately owned vehicles owned by households in the traffic analysis zone by place of residence}
\item{ft_workers}{Number of people employed full-time in the traffic analysis zone by place of residence}
\item{pt_workers}{Number of people employed part-time in the traffic analysis zone by place of residence}
\item{median_age}{Median age of population in the traffic analysis zone}
\item{House}{Number of households in the traffic analysis zone who live in dwelling of type House}
\item{Apartment}{Number of households in the traffic analysis zone who live in dwelling of type Apartment}
\item{Townhouse}{Number of households in the traffic analysis zone who live in dwelling of type Townhouse}
\item{total_dwellings}{Total number of households in the traffic analysis zone}
\item{median_household_size}{Median number of people per households in the traffic analysis zone}
\item{Inc_less_15k}{Number of households in the traffic analysis zone with before-tax household income from all sources $14,999 or less}
\item{Inc_15k_to_40k}{Number of households in the traffic analysis zone with before-tax household income from all sources between $15,000 and $39,999}
\item{Inc_40k_to_60k}{Number of households in the traffic analysis zone with before-tax household income from all sources between $40,000 and $59,999}
\item{Inc_60k_to_100k}{Number of households in the traffic analysis zone with before-tax household income from all sources between $60,000 and $99,999}
\item{Inc_100k_to_125k}{Number of households in the traffic analysis zone with before-tax household income from all sources between $100,000 and $124,999}
\item{Inc_15k_to_40k}{Number of households in the traffic analysis zone with before-tax household income from all sources $125,000 and greater}
\item{Inc_decline_dk}{Number of households in the traffic analysis zone who declined to respond or did not know their before-tax household income from all sources}
}
}
\source{
Transportation Tomorrow Survey (\url{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction})

Data Guide available (\url{http://dmg.utoronto.ca/pdf/tts/2016/2016TTS_DataGuide.pdf})
}
\usage{
data(hamilton_taz_ph)
}
\description{
A table with statistics about persons and households by traffic analysis zones in
the Hamilton CMA. The zoning system is GTA06. This file is a portion of the system
used by the Transportation Tomorrow Survey in the Greater Toronto and Hamilton Area.
}
\examples{
 data(hamilton_taz_ph)
 summary(hamilton_taz_ph)
}
\keyword{datasets}
\keyword{transportation}
