% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_taz_travel}
\alias{hamilton_taz_travel}
\title{Travel-related statistics for traffic analysis zones in the Hamilton CMA.}
\format{
A simple features object with 297 rows and 17 variables:
\describe{
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{work_trips_produced}{Number of trips produced by residents of the traffic analysis zone for work purposes (first work trip of day)}
\item{work_trips_attracted}{Number of trips attracted by employers in the traffic analysis zone for work purposes (first work trip of day)}
\item{shop_trips_produced}{Number of trips produced by residents of the traffic analysis zone for shopping purposes (Marketin/Shopping)}
\item{shop_trips_attracted}{Number of trips attracted by businesses in the traffic analysis zone for shopping purposes (Marketin/Shopping)}
\item{Walk}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Walk}
\item{Cycle}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Cycle}
\item{Walk}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Walk}
\item{Auto_driver}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Auto Driver}
\item{Auto_passenger}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Auto Passenger}
\item{Motorcycle}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Motorcycle}
\item{Taxi_passenger}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Taxi Passenger}
\item{Paid_rideshare}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Paid Rideshare (Uber, Lyft, DriverHer, or other paid rideshare app)}
\item{School_bus}{Number of trips made by residents of the traffic analysis zone for all purposes by mode Schoolbus}
\item{Transit_excluding_GO_rail}{Number of trips made by residents of the traffic analysis zone for all purposes by Transit excluding GO rail}
\item{GO_rail_and_transit}{Number of trips made by residents of the traffic analysis zone for all purposes by a combination of GO rail and local transit services}
\item{GO_rail}{Number of trips made by residents of the traffic analysis zone for all purposes by GO rail exclusively}
\item{Other}{Number of trips made by residents of the traffic analysis zone for all purposes by any other mode}
}
}
\source{
Transportation Tomorrow Survey (\url{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction})

Data Guide available (\url{http://dmg.utoronto.ca/pdf/tts/2016/2016TTS_DataGuide.pdf})
}
\usage{
data(hamilton_taz_travel)
}
\description{
A table with statistics about trips and travel by traffic analysis zones in
the Hamilton CMA. The zoning system is GTA06. This file is a portion of the system
used by the Transportation Tomorrow Survey in the Greater Toronto and Hamilton Area.
}
\examples{
 data(hamilton_taz_travel)
 summary(hamilton_taz_travel)
}
\keyword{datasets}
\keyword{transportation}
