% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_trips_mode}
\alias{hamilton_trips_mode}
\title{Trips by mode in the Hamilton CMA.}
\format{
A data frame with 297 rows and 5 variables:
\describe{
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{Cycle}{Number of trips that originated in the traffic analysis by mode: Cycle}
\item{Auto_driver}{Number of trips that originated in the traffic analysis by mode: Auto_driver}
\item{Auto_passenger}{Number of trips that originated in the traffic analysis by mode: Auto_passenger}
\item{Walk}{Number of trips that originated in the traffic analysis by mode: Walk}
}
}
\source{
Transportation Tomorrow Survey (http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction)
}
\usage{
data(hamilton_trips_mode)
}
\description{
A table with information from the Transportation Tomorrow Survey regarding trips
by mode by traffic analysis zone. These data are for 2016.
}
\examples{
 data(hamilton_trips_mode)
 summary(hamilton_trips_mode)
}
\keyword{datasets}
\keyword{transportation}
