% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_trips_purpose}
\alias{hamilton_trips_purpose}
\title{Trips by purpose in the Hamilton CMA.}
\format{
A data frame with 297 rows and 26 variables:
\describe{
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{Unknown}{Number of trips that originated in the traffic analysis zone for purpose: Unknown}
\item{Subsequent_School}{Number of trips that originated in the traffic analysis zone for purpose: Subsequent_School}
\item{Daycare}{Number of trips that originated in the traffic analysis zone for purpose: Daycare}
\item{Facilitate_passenger}{Number of trips that originated in the traffic analysis zone for purpose: Facilitate_passenger}
\item{Home}{Number of trips that originated in the traffic analysis zone for purpose: Home}
\item{Market_Shop}{Number of trips that originated in the traffic analysis zone for purpose: Market_Shop}
\item{Other}{Number of trips that originated in the traffic analysis zone for purpose: Other}
\item{Subsequent_Work}{Number of trips that originated in the traffic analysis zone for purpose: Subsequent_Work}
\item{School}{Number of trips that originated in the traffic analysis zone for purpose: School}
\item{Work}{Number of trips that originated in the traffic analysis zone for purpose: Work}
}
}
\source{
Transportation Tomorrow Survey (http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction)
}
\usage{
data(hamilton_trips_purpose)
}
\description{
A table with information from the Transportation Tomorrow Survey regarding trips
by purpose by traffic analysis zone. These data are for 2016.
}
\examples{
 data(hamilton_trips_purpose)
 summary(hamilton_trips_purpose)
}
\keyword{datasets}
\keyword{transportation}
