% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{od_full}
\alias{od_full}
\title{Origin-destination matrix for the Hamilton CMA.}
\format{
A simple features with 87,912 rows and 10 variables:
\describe{
\item{Origin}{Unique identifier of zone of origin of trips in the GTA06 system}
\item{Destination}{Unique identifier of zone of destination of trips in the GTA06 system}
\item{x_origin}{X coordinate of the centroid of the zone of origin of trips}
\item{y_origin}{Y coordinate of the centroid of the zone of origin of trips}
\item{x_destination}{X coordinate of the centroid of the zone of destination of trips}
\item{y_destination}{Y coordinate of the centroid of the zone of destination of trips}
\item{PairID}{Unique identifier for origin-destination pair}
\item{Trips_work}{Number of trips for work purposes between origin-destination}
\item{Trips_shop}{Number of trips for shop purposes between origin-destination}
\item{geometry}{Geometry column}
}
}
\source{
Transportation Tomorrow Survey (\url{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction})

Data Guide available (\url{http://dmg.utoronto.ca/pdf/tts/2016/2016TTS_DataGuide.pdf})
}
\usage{
data(od_full)
}
\description{
A table with information about trips that originated in the Hamilton CMA and concluded in
the Hamilton CMA. This excludes trips that originated outside of the Hamilton CMA even if
they concluded in the Hamilton CMA. Similarly, it excludes trips that originated in the
Hamilton CMA but concluded outside of the Hamilton CMA. The zoning system is GTA06. This
file is a portion of the system used by the Transportation Tomorrow Survey in the Greater
Toronto and Hamilton Area. Origins and destinations are connected using straight line distances.
The object is projected as UTM Zone 17N.
}
\examples{
 data(od_full)
 summary(od_full)
}
\keyword{datasets}
\keyword{transportation}
