% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{travel_cost_car}
\alias{travel_cost_car}
\title{Travel cost by car in the Hamilton CMA.}
\format{
A data frame with 297 rows and 6 variables:
\describe{
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{from}{Address in the traffic analysis zone used for routing}
\item{to}{Address of destination}
\item{km}{Distance of trip in km}
\item{minutes}{Duration of trip in minutes}
\item{hours}{Duration of trip in hours}
}
}
\source{
Estimated travel distances and durations obtained from Google Maps Routing API
}
\usage{
data(travel_cost_car)
}
\description{
A table with information about the estimated cost of travel from each traffic analysis zone
to downtown Hamilton (Jackson Square).
}
\examples{
 data(travel_cost_car)
 summary(travel_cost_car)
}
\keyword{datasets}
\keyword{transportation}
