% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{package}
\name{envsocty3LT3-package}
\alias{envsocty3LT3}
\alias{envsocty3LT3-package}
\title{envsocty3LT3: A Course Package for Transportation Geography}
\description{
Course package for ENVSOCTY 3LT3 Transportation Geography.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/envsocty3LT3}
  \item \url{https://paezha.github.io/envsocty3LT3/}
  \item Report bugs at \url{https://github.com/paezha/envsocty3LT3/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Shaila Jamal \email{jamals16@mcmaster.ca} (\href{https://orcid.org/0000-0002-3628-4346}{ORCID})
  \item Elise Desjardins \email{desjae@mcmaster.ca } (\href{https://orcid.org/0000-0002-3350-4534}{ORCID})
}

}
\keyword{internal}
