% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envsocty3LT3-package.R
\docType{data}
\name{hamilton_taz}
\alias{hamilton_taz}
\title{Traffic analysis zones in the Hamilton CMA.}
\format{
A simple features object with 297 rows and 8 variables:
\describe{
\item{ID}{Unique identifier for traffic analysis zones}
\item{Area}{Surface area of zone in sq.km}
\item{GTA06}{Unique identifier of zone in the GTA06 system}
\item{geometry}{Geometry column}
}
}
\source{
Transportation Tomorrow Survey (http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction)
}
\usage{
data(hamilton_taz)
}
\description{
A table with the traffic analysis zones in the Hamilton CMA, using the GTA06 system.
Traffic analysis zones are zoning systems designed for transportation studies. This
file is a portion of the system used by the Transportation Tomorrow Survey in the
Greater Toronto and Hamilton Area. The object is projected as UTM Zone 17N.
}
\examples{
 data(hamilton_taz)
 summary(hamilton_taz)
}
\keyword{datasets}
\keyword{transportation}
